Refactor testfs code for clarity
