+++ lib/testfs/index.ts
@@ -1,0 +1,2 @@
+export * from './testfs';
+export * from './types';
