+++ tests/hooks.spec.ts
@@ -1,4 +1,4 @@
-import { testfs, MochaPodConfig } from '~/mocha-pod';
+import { testfs, MochaPod } from '~/mocha-pod';
 import { promises as fs } from 'fs';
 
 import { expect } from './chai';
@@ -11,7 +11,7 @@ describe('hooks: integration tests', function () {
 	it('global testfs configuration should be used', async () => {
 		// Here we just check that the filesystem entry was set. Hooks
 		// have already ran so we cannot modify config at this stage
-		const conf = await MochaPodConfig();
+		const conf = await MochaPod.Config();
 		expect(conf.testfs.filesystem).to.not.be.undefined;
 
 		await expect(
@@ -19,7 +19,7 @@ describe('hooks: integration tests', function () {
 			'global testfs files should not exist before testfs.setup()',
 		).to.be.rejected;
 
-		const tmp = await testfs().setup();
+		const tmp = await testfs().enable();
 
 		// The file contents should match whatever is in config
 		expect(await fs.readFile('/etc/unused.conf', 'utf-8')).to.equal(
